[DRIV-103] - Removed RedisClientHelper interface
